Move Buffer to shared
